requirements.txt
alembic
bcrypt
cachetools
certifi
cffi
chardet
click
cryptography
cssselect
cssutils
dnspython
ecdsa
email-validator
emails
fastapi-sqlalchemy
fastapi
greenlet
h11
idna
lxml
mako
markupsafe
passlib[bcrypt]
premailer
psycopg2-binary
psycopg2
pyasn1
pycparser
pydantic[email]
python-dateutil
python-dotenv
python-editor
python-jose[cryptography]
python-multipart
requests
rsa
six
sqlalchemy
starlette
typing-extensions
urllib3
uvicorn
